extract common functions from Source Connector and Task
